Fix temporarily a bug causing crashes when aligning coverage map
The temporary solution is to set the w of orientation as 1, until
slam is able to send as correct orientation(for now z and w are not
sent)

Sensor coverage node is run with a debugger for now. It seems there
are now problems or crashes for now, if you find anything please
report it!